refactor(frontend): move borrow into it's own store - enable debugging with devtools - enable better typing
