+++ packages/frontend/components/Borrow/Borrow.tsx
@@ -23,14 +23,15 @@ import LoadingButton from "@mui/lab/LoadingButton"
 import { useHistory } from "../../store/history.store"
 import NetworkIcon from "../NetworkIcon"
 import { chainName } from "../../helpers/chainName"
+import { useBorrow } from "../../store/borrow.store"
 
 export default function Borrow() {
   const address = useStore((state) => state.address)
   const walletChain = useStore((state) => state.chain)
   const changeChain = useStore((state) => state.changeChain)
-  const updateBalance = useStore((state) => state.updateBalances)
-  const updateVault = useStore((state) => state.updateVault)
-  const updateAllowance = useStore((state) => state.updateAllowance)
+  const updateBalance = useBorrow((state) => state.updateBalances)
+  const updateVault = useBorrow((state) => state.updateVault)
+  const updateAllowance = useBorrow((state) => state.updateAllowance)
   useEffect(() => {
     if (address) {
       updateBalance("collateral")
@@ -41,47 +42,49 @@ export default function Borrow() {
   }, [address, updateBalance, updateAllowance, updateVault])
 
   const login = useStore((state) => state.login)
-  const formType = useStore((state) => state.formType)
+  const formType = useBorrow((state) => state.formType)
 
-  const collateral = useStore((state) => state.position.collateral)
-  const collateralInput = useStore((state) => state.collateralInput)
+  const collateral = useBorrow((state) => state.position.collateral)
+  const collateralInput = useBorrow((state) => state.collateralInput)
   const collateralAmount = parseFloat(collateralInput)
-  const collateralChainId = useStore((state) => state.collateralChainId)
-  const collateralAllowance = useStore((state) => state.collateralAllowance)
+  const collateralChainId = useBorrow((state) => state.collateralChainId)
+  const collateralAllowance = useBorrow((state) => state.collateralAllowance)
 
-  const debtInput = useStore((state) => state.debtInput)
+  const debtInput = useBorrow((state) => state.debtInput)
   const debtAmount = parseFloat(debtInput)
-  const debtChainId = useStore((state) => state.debtChainId)
+  const debtChainId = useBorrow((state) => state.debtChainId)
 
-  const vaultChainId = useStore((state) => state.position.vault?.chainId)
+  const vaultChainId = useBorrow((state) => state.position.vault?.chainId)
 
-  const changeBorrowChain = useStore((state) => state.changeBorrowChain)
-  const changeCollateralChain = useStore((state) => state.changeCollateralChain)
+  const changeBorrowChain = useBorrow((state) => state.changeBorrowChain)
+  const changeCollateralChain = useBorrow(
+    (state) => state.changeCollateralChain
+  )
 
   // TODO: refacto with a "status" in store (i.e status = "editing, approving, signing, borrowing...") ?
   const [showApprovalModal, setShowApprovalModal] = useState(false)
 
-  const balance = useStore(
+  const balance = useBorrow(
     (state) => state.collateralBalances[state.position.collateral.token.symbol]
   )
 
-  const updateTokenPrice = useStore((state) => state.updateTokenPrice)
+  const updateTokenPrice = useBorrow((state) => state.updateTokenPrice)
   useEffect(() => {
     updateTokenPrice("collateral")
     updateTokenPrice("debt")
   }, [updateTokenPrice])
 
-  const ltv = useStore((state) => state.position.ltv)
-  const ltvMax = useStore((state) => state.position.ltvMax)
+  const ltv = useBorrow((state) => state.position.ltv)
+  const ltvMax = useBorrow((state) => state.position.ltvMax)
 
-  const signAndBorrow = useStore((state) => state.signAndBorrow)
-  const isSigning = useStore((state) => state.isSigning)
-  const isBorrowing = useStore((state) => state.isBorrowing)
+  const signAndBorrow = useBorrow((state) => state.signAndBorrow)
+  const isSigning = useBorrow((state) => state.isSigning)
+  const isBorrowing = useBorrow((state) => state.isBorrowing)
 
   const currentTxHash = useHistory((state) => state.inModal)
   const closeModal = useHistory((state) => state.closeModal)
-  const metaStatus = useStore((state) => state.transactionMeta.status)
-  const availableVaultStatus = useStore((state) => state.availableVaultsStatus)
+  const metaStatus = useBorrow((state) => state.transactionMeta.status)
+  const availableVaultStatus = useBorrow((state) => state.availableVaultsStatus)
 
   let button: ReactNode
   if (!address) {
