+++ packages/frontend/components/Borrow/TransactionSummary.tsx
@@ -18,27 +18,27 @@ import KeyboardArrowDownIcon from "@mui/icons-material/KeyboardArrowDown"
 
 import LTVProgressBar from "./LTVProgressBar"
 import ClickableTooltip from "../Layout/ClickableTooltip"
-import { useStore } from "../../store"
 import { formatUnits } from "ethers/lib/utils"
 import ProviderIcon from "../ProviderIcon"
+import { useBorrow } from "../../store/borrow.store"
 
 export default function TransactionSummary() {
   const { palette } = useTheme()
 
-  const ltv = useStore((state) => state.position.ltv)
-  const ltvMax = useStore((state) => state.position.ltvMax)
-  const liquidationPrice = useStore((state) => state.position.liquidationPrice)
-  const liquidationDiff = useStore((state) => state.position.liquidationDiff)
+  const ltv = useBorrow((state) => state.position.ltv)
+  const ltvMax = useBorrow((state) => state.position.ltvMax)
+  const liquidationPrice = useBorrow((state) => state.position.liquidationPrice)
+  const liquidationDiff = useBorrow((state) => state.position.liquidationDiff)
 
-  const collateral = useStore((state) => state.position.collateral)
-  const collateralInput = useStore((state) => state.collateralInput)
+  const collateral = useBorrow((state) => state.position.collateral)
+  const collateralInput = useBorrow((state) => state.collateralInput)
   const collateralAmount = parseFloat(collateralInput)
 
-  const debt = useStore((state) => state.position.debt)
-  const debtInput = useStore((state) => state.debtInput)
+  const debt = useBorrow((state) => state.position.debt)
+  const debtInput = useBorrow((state) => state.debtInput)
   const debtAmount = parseFloat(debtInput)
 
-  const providers = useStore((state) => state.position.providers)
+  const providers = useBorrow((state) => state.position.providers)
 
   const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null)
   const isOpen = Boolean(anchorEl)
