+++ packages/frontend/store/history.store.ts
@@ -8,11 +8,11 @@ import {
 import produce from "immer"
 import create from "zustand"
 import { devtools, persist } from "zustand/middleware"
-import { useStore } from "."
 import { sdk } from "./auth.slice"
 import ethers from "ethers"
 import { useSnack } from "./snackbar.store"
 import { formatUnits } from "ethers/lib/utils"
+import { useBorrow } from "./borrow.store"
 
 export type HistoryStore = HistoryState & HistoryActions
 
@@ -176,11 +176,11 @@ export const useHistory = create<HistoryStore>()(
               )
               console.debug(step.step, txHash)
               if (step.step === RoutingStep.DEPOSIT) {
-                useStore.getState().updateBalances("collateral")
-                useStore.getState().updateAllowance()
+                useBorrow.getState().updateBalances("collateral")
+                useBorrow.getState().updateAllowance()
               }
               if (step.step === RoutingStep.BORROW) {
-                useStore.getState().updateBalances("collateral")
+                useBorrow.getState().updateBalances("collateral")
               }
               // TODO: can we have error ? if yes mark the tx as failed. Design ? Retry ?
             }
