+++ packages/frontend/store/index.ts
@@ -1,13 +1,12 @@
 import create from "zustand"
 import { createAuthSlice, AuthStore } from "./auth.slice"
-import { createTransactionSlice, BorrowStore } from "./borrow.slice"
 import { devtools } from "zustand/middleware"
 
-export const useStore = create<AuthStore & BorrowStore>()(
+export const useStore = create<AuthStore>()(
   // devtools(
   (...a) => ({
     ...createAuthSlice(...a),
-    ...createTransactionSlice(...a),
+    // ...createTransactionSlice(...a),
   })
   // {
   // enabled: process.env.NODE_ENV === "development",
